Refactoring, moving controllers to other package
