+++ src/userinterface/controllers/ManagerController.java
@@ -1,25 +1,13 @@
-package userinterface.view;
+package userinterface.controllers;
 
-import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
-import javafx.beans.value.ObservableValue;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
+
 import javafx.fxml.Initializable;
-import javafx.scene.Cursor;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.*;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
-import javafx.stage.Modality;
 import javafx.stage.Stage;
 import userinterface.MainApp;
-import userinterface.utils.SplitPaneDividerSlider;
 
 public class ManagerController extends Controller implements Initializable {
     private MainApp mainApp;
