+++ src/userinterface/controllers/PassengerInfoController.java
@@ -1,7 +1,6 @@
-package userinterface.view;
+package userinterface.controllers;
 
 import javafx.beans.value.ObservableValue;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
