+++ src/userinterface/controllers/PassengersDeletingController.java
@@ -1,4 +1,4 @@
-package userinterface.view;
+package userinterface.controllers;
 
 import javafx.fxml.FXML;
 import javafx.stage.Stage;
