+++ src/userinterface/view/ManagerScene.fxml
@@ -16,7 +16,7 @@
 <?import javafx.scene.control.ChoiceBox?>
 <?import java.lang.String?>
 
-<BorderPane fx:id="managerScene" prefHeight="840.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userinterface.view.ManagerController">
+<BorderPane fx:id="managerScene" prefHeight="840.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userinterface.controllers.ManagerController">
     <center>
         <TabPane  tabClosingPolicy="UNAVAILABLE" >
             <tabs>
