+++ src/userinterface/view/PassengerLayout.fxml
@@ -20,7 +20,7 @@
 <?import javafx.scene.control.ChoiceBox?>
 <?import javafx.scene.control.cell.PropertyValueFactory?>
 
-<AnchorPane prefHeight="840.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/8.0.76-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userinterface.view.PassengerInfoController" >
+<AnchorPane prefHeight="840.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/8.0.76-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userinterface.controllers.PassengerInfoController" >
     <children>
         <StackPane prefHeight="150.0" prefWidth="200.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
             <children>
